pull confirmation functions into separate file
